trading-bot/requirements.txt
MetaTrader5==5.0.4424
numpy
pandas
pytz
websocket-client
requests
rel
zmq
